+++ test/diff.test.js
@@ -176,3 +176,35 @@ test('diff-timezone equal times', (t) => {
   t.equal(diff.seconds, 0, 'same-sec')
   t.end()
 })
+
+test('i18n', (t) => {
+  let start = spacetime('Dec 25th 2021')
+  let end = spacetime('Feb 2nd 2022')
+
+  let translationValues = {
+    units: {
+      secondWord: 'segundo',
+      secondWordPlural: 'segundos',
+      minuteWord: 'minuto',
+      minuteWordPlural: 'minutos',
+      hourWord: 'hora',
+      hourWordPlural: 'horas',
+      dayWord: 'dia',
+      dayWordPlural: 'dias',
+      monthWord: 'mes',
+      monthWordPlural: 'meses',
+      yearWord: 'año',
+      yearWordPlural: 'años'
+    }
+  }
+
+  start.i18n(translationValues)
+  end.i18n(translationValues)
+
+  let diff = start.since(end).diff
+  t.equal(diff.days, -8, 'same-day')
+  t.equal(diff.hours, 0, 'hour diff')
+  t.equal(diff.minutes, 0, 'same-min')
+  t.equal(diff.seconds, 0, 'same-sec')
+  t.end()
+})